LayoutTests/fast/css-grid-layout/non-grid-columns-rows-get-set.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<link href="resources/grid.css" rel="stylesheet">
<style>
.gridWithNone {
    -webkit-grid-template-columns: none;
    -webkit-grid-template-rows: none;
}
.gridWithFixed {
    -webkit-grid-template-columns: 10px;
    -webkit-grid-template-rows: 15px;
}
.gridWithPercent {
    -webkit-grid-template-columns: 53%;
    -webkit-grid-template-rows: 27%;
}
.gridWithAuto {
    -webkit-grid-template-columns: auto;
    -webkit-grid-template-rows: auto;
}
.gridWithEM {
    -webkit-grid-template-columns: 10em;
    -webkit-grid-template-rows: 15em;
    font: 10px Ahem;
}
.gridWithViewPortPercentage {
    -webkit-grid-template-columns: 8vw;
    -webkit-grid-template-rows: 10vh;
}
.gridWithFitContent {
    -webkit-grid-template-columns: -webkit-fit-content;
    -webkit-grid-template-rows: -webkit-fit-content;
}
.gridWithFitAvailable {
    -webkit-grid-template-columns: -webkit-fit-available;
    -webkit-grid-template-rows: -webkit-fit-available;
}
.gridWithMinMax {
    -webkit-grid-template-columns: minmax(10%, 15px);
    -webkit-grid-template-rows: minmax(20px, 50%);
}
.gridWithMinContent {
    -webkit-grid-template-columns: -webkit-min-content;
    -webkit-grid-template-rows: -webkit-min-content;
}
.gridWithMaxContent {
    -webkit-grid-template-columns: -webkit-max-content;
    -webkit-grid-template-rows: -webkit-max-content;
}
.gridWithFraction {
    -webkit-grid-template-columns: 1fr;
    -webkit-grid-template-rows: 2fr;
}
</style>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div class="gridWithNone" id="gridWithNoneElement"></div>
<div class="gridWithFixed" id="gridWithFixedElement"></div>
<div class="gridWithPercent" id="gridWithPercentElement"></div>
<div class="gridWithAuto" id="gridWithAutoElement"></div>
<div class="gridWithEM" id="gridWithEMElement"></div>
<div class="gridWithViewPortPercentage" id="gridWithViewPortPercentageElement"></div>
<div class="gridWithFitContent" id="gridWithFitContentElement"></div>
<div class="gridWithFitAvailable" id="gridWithFitAvailableElement"></div>
<div class="gridWithMinMax" id="gridWithMinMax"></div>
<div class="gridWithMinContent" id="gridWithMinContent"></div>
<div class="gridWithMaxContent" id="gridWithMaxContent"></div>
<div class="gridWithFraction" id="gridWithFraction"></div>
<script src="resources/grid-definitions-parsing-utils.js"></script>
<script src="resources/non-grid-columns-rows-get-set.js"></script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
